+comment

+++ test_haltec/delete_user.php
@@ -1,10 +1,10 @@
 <?php
-include 'conn.php';
+include 'conn.php'; //ngambil koneksi
 
-$userId = $_GET['id'] ?? '';
-$deleteUserQuery = "DELETE FROM user WHERE id = $userId";
+$userId = $_GET['id'] ?? ''; //ngambil id dari url link dengan method get
+$deleteUserQuery = "DELETE FROM user WHERE id = $userId"; //query delete
 if (mysqli_query($conn, $deleteUserQuery)) {
-    $deleteRoleQuery = "DELETE FROM user_roles WHERE id_user = $userId";
+    $deleteRoleQuery = "DELETE FROM user_roles WHERE id_user = $userId"; //query delete roles
     if (mysqli_query($conn, $deleteRoleQuery)) {
         header("Location: dashboard.php?roles=admin");
         exit();
@@ -14,4 +14,5 @@ if (mysqli_query($conn, $deleteUserQuery)) {
 } else {
     echo "Error deleting user: " . mysqli_error($conn);
 }
+
 ?>
